+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./requests";
 export * from "./retry";
 export * from "./sleep";
 export * from "./ton";
+export * from "./registryUtils";
